requirements.txt
celery==4.4.2
Django==3.0.4
django-redis==4.11.0
djangorestframework==3.11.0
graphene-django
psycopg2==2.8.4
redis==3.4.1